First Version of PreviewPhotoTaker.
First, the user has to give the name of the object, that will be photographed. This name will be the name of the folder, where the taken images will be stored in the "PhotosForResearch" folder on the SD-card. The image names are numbers, they are always stored in ascending order.
If the user touches the preview, the preview image will be saved.